# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makers Community API",
        "description": "API complète pour la plateforme communautaire Makers Community avec authentification, posts, commentaires, ratings et modération automatique. Cette API permet de gérer une communauté de makers avec toutes les fonctionnalités nécessaires.",
        "contact": {
            "name": "Équipe Makers Community",
            "email": "dev@makerscomm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.makerscomm.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Dashboard administrateur",
                "description": "Récupérer les statistiques générales de la plateforme pour le dashboard admin",
                "operationId": "7cfa73c7f4bc906c66043dac50a3d549",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminDashboard"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard chargé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - Admin uniquement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vous devez être administrateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister tous les utilisateurs",
                "description": "Récupérer la liste paginée des utilisateurs avec filtres et statistiques",
                "operationId": "7a629ec3ab4a321032bfa425c844d09d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filtrer par rôle",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "user",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher par nom ou email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Utilisateur",
                "description": "Modèle utilisateur de base",
                "properties": {
                    "id": {
                        "description": "Identifiant unique",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom complet",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "birthday": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date",
                        "example": "1995-05-15"
                    },
                    "role": {
                        "description": "Rôle utilisateur",
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    },
                    "profile_pic": {
                        "description": "Photo de profil",
                        "type": "string",
                        "example": "profile_123.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif du compte",
                        "type": "boolean",
                        "example": true
                    },
                    "email_verified_at": {
                        "description": "Date de vérification email",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "AdminDashboard": {
                "title": "Dashboard Admin",
                "properties": {
                    "users": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 1250
                            },
                            "active": {
                                "type": "integer",
                                "example": 1180
                            },
                            "inactive": {
                                "type": "integer",
                                "example": 70
                            },
                            "admins": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    },
                    "posts": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 450
                            },
                            "this_month": {
                                "type": "integer",
                                "example": 23
                            },
                            "average_rating": {
                                "type": "number",
                                "format": "float",
                                "example": 4.2
                            }
                        },
                        "type": "object"
                    },
                    "reports": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 12
                            },
                            "pending": {
                                "type": "integer",
                                "example": 3
                            },
                            "resolved": {
                                "type": "integer",
                                "example": 9
                            }
                        },
                        "type": "object"
                    },
                    "activity": {
                        "properties": {
                            "comments_today": {
                                "type": "integer",
                                "example": 15
                            },
                            "ratings_today": {
                                "type": "integer",
                                "example": 8
                            },
                            "new_users_today": {
                                "type": "integer",
                                "example": 3
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token d'authentification Laravel Sanctum. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}